+++ MatrixSDK/Crypto/Data/MXDeviceListOperationsPool.m
@@ -115,14 +115,19 @@
             if (crossSigningKeys)
             {
                 NSLog(@"[MXDeviceListOperationsPool] doKeyDownloadForUsers: Got cross-signing keys for %@: %@", userId, crossSigningKeys);
+                
+                MXCrossSigningInfo *storedCrossSigningKeys = [self->crypto.store crossSigningKeysForUser:userId];
+                
+                // Use current trust level
+                MXUserTrustLevel *oldTrustLevel = storedCrossSigningKeys.trustLevel;
+                [crossSigningKeys setTrustLevel:oldTrustLevel];
 
                 // Compute trust on this user
                 // Note this overwrites the previous value
                 BOOL isCrossSigningVerified = [self->crypto.crossSigning isUserWithCrossSigningKeysVerified:crossSigningKeys];
-                
-                BOOL wasLocallyVerified = [self->crypto.store crossSigningKeysForUser:userId].trustLevel.isLocallyVerified;
                 MXUserTrustLevel *newTrustLevel = [MXUserTrustLevel trustLevelWithCrossSigningVerified:isCrossSigningVerified
-                                                                                       locallyVerified:wasLocallyVerified];
+                                                                                       locallyVerified:oldTrustLevel.isLocallyVerified];
+                
                 [crossSigningKeys updateTrustLevel:newTrustLevel];
                 
                 // Note that keys which aren't in the response will be removed from the store
@@ -167,11 +172,17 @@
                         // So, transfer its previous value
                         previousLocalState = previouslyStoredDeviceKeys.trustLevel.localVerificationStatus;
                     }
-
+                    
+                    // Use current trust level
+                    MXDeviceTrustLevel *oldTrustLevel = [MXDeviceTrustLevel trustLevelWithLocalVerificationStatus:previousLocalState
+                                                                                             crossSigningVerified:previouslyStoredDeviceKeys.trustLevel.isCrossSigningVerified];
+                    [mutabledevices[deviceId] setTrustLevel:oldTrustLevel];
+                    
+                    
                     BOOL crossSigningVerified = [self->crypto.crossSigning isDeviceVerified:mutabledevices[deviceId]];
                     MXDeviceTrustLevel *trustLevel = [MXDeviceTrustLevel trustLevelWithLocalVerificationStatus:previousLocalState
                                                                                           crossSigningVerified:crossSigningVerified];
-
+                    
                     [mutabledevices[deviceId] updateTrustLevel:trustLevel];
                 }
 
@@ -185,7 +196,6 @@
                     // Note that devices which aren't in the response will be removed from the store
                     [self->crypto.store storeDevicesForUser:userId devices:mutabledevices];
                 }
-
             }
         }